+++ feature/library/src/main/java/com/into/websoso/feature/library/filter/component/LibraryFilterBottomSheetButtons.kt
@@ -1,4 +1,4 @@
-package com.into.websoso.feature.filter.component
+package com.into.websoso.feature.library.filter.component
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
