+++ feature/library/src/main/java/com/into/websoso/feature/library/filter/component/LibraryFilterBottomSheetHeader.kt
@@ -1,4 +1,4 @@
-package com.into.websoso.feature.filter.component
+package com.into.websoso.feature.library.filter.component
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
